tblog/requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-recaptcha==2.0.6
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.2.0
wrapt==1.12.1
